port: Misc mixin fixes

+++ common/src/main/java/com/wynntils/mc/event/MobEffectEvent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © Wynntils 2022-2023.
+ * Copyright © Wynntils 2022-2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.mc.event;
@@ -26,16 +26,16 @@ public abstract class MobEffectEvent extends Event {
     }
 
     public static final class Update extends MobEffectEvent {
-        private final byte effectAmplifier;
+        private final int effectAmplifier;
         private final int effectDurationTicks;
 
-        public Update(Entity entity, MobEffect effect, byte effectAmplifier, int effectDurationTicks) {
+        public Update(Entity entity, MobEffect effect, int effectAmplifier, int effectDurationTicks) {
             super(entity, effect);
             this.effectAmplifier = effectAmplifier;
             this.effectDurationTicks = effectDurationTicks;
         }
 
-        public byte getEffectAmplifier() {
+        public int getEffectAmplifier() {
             return effectAmplifier;
         }
